fix: used correct initial value for s_miniMedian

+++ contracts/libraries/PanopticMath.sol
@@ -31,10 +31,10 @@ library PanopticMath {
     uint256 internal constant UPPER_120BITS_MASK =
         0xFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF00000000000000000000000000000000;
 
-    uint256 constant EMA_PERIOD_10MINS = 600; // 600 seconds
-    uint256 constant EMA_PERIOD_1H = 3600; // 600 seconds
-    uint256 constant EMA_PERIOD_8H = 28800; // 600 seconds
-    uint256 constant EMA_PERIOD_1D = 86400; // 600 seconds
+    int256 constant EMA_PERIOD_10MINS = 600; // 600 seconds
+    int256 constant EMA_PERIOD_1H = 3600; // 600 seconds
+    int256 constant EMA_PERIOD_8H = 28800; // 600 seconds
+    int256 constant EMA_PERIOD_1D = 86400; // 600 seconds
 
     /*//////////////////////////////////////////////////////////////
                               UTILITIES
@@ -282,12 +282,12 @@ library PanopticMath {
 
             uint256 currentEpoch;
             bool differentEpoch;
-            uint256 timeDelta;
+            int256 timeDelta;
             {
                 currentEpoch = (block.timestamp >> 6) & 0xFFFFFF; // mod 2**24
                 uint256 recordedEpoch = medianData >> 232;
                 differentEpoch = currentEpoch != recordedEpoch;
-                timeDelta = (currentEpoch - recordedEpoch) * 64;
+                timeDelta = int256((currentEpoch - recordedEpoch) * 64);
             }
             // only proceed if last entry is in a different epoch (takes care of looping edge case in a way that ">" doesn't)
             if (differentEpoch) {
@@ -299,13 +299,14 @@ library PanopticMath {
 
                 int24 clampedTick = clampTick(lastObservedTick, medianData);
 
+                /*
                 console2.log(
                     "dd",
                     int24(uint24(medianData >> 96)) +
-                        ((clampedTick - int24(uint24(medianData >> 96))) * timeDelta) /
+                        ((clampedTick - int24(uint24(medianData >> 96))) * int256(timeDelta)) /
                         EMA_PERIOD_10MINS
                 );
-
+                */
                 updatedMedianData = insertObservation(medianData, clampedTick, currentEpoch);
             }
         }
